server/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.0
Flask-Login==0.6.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
passlib==1.7.4
types-passlib==1.7.7.13
flake8==6.1.0
flask-cors==4.0.0
Werkzeug==2.0.2
SQLAlchemy==1.4.26
bcrypt==4.0.1
sendgrid==6.10.0
scikit-learn==1.3.1
requests==2.31.0
dill==0.3.7
gevent==23.9.1
aiohttp==3.8.4
asyncio==3.4.3
waitress==2.1.2
gunicorn==20.1.0